Move tests to appropriate file
