+++ src/test/java/org/opentripplanner/openstreetmap/model/OSMWayTest.java
@@ -130,215 +130,6 @@ public class OSMWayTest {
     assertTrue(way.isOpposableCycleway());
   }
 
-  /**
-   * Tests if cars can drive on unclassified highways with bicycleDesignated
-   * <p>
-   * Check for bug #1878 and PR #1880
-   */
-  @Test
-  void testCarPermission() {
-    OSMWay way = new OSMWay();
-    way.addTag("highway", "unclassified");
-
-    var permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-
-    way.addTag("bicycle", "designated");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-  }
-
-  /**
-   * Tests that motorcar/bicycle/foot private don't add permissions but yes add permission if access
-   * is no
-   */
-  @Test
-  void testMotorCarTagAllowedPermissions() {
-    OSMWay way = new OSMWay();
-    way.addTag("highway", "residential");
-    var permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-
-    way.addTag("access", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allowsNothing());
-
-    way.addTag("motorcar", "private");
-    way.addTag("bicycle", "private");
-    way.addTag("foot", "private");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allowsNothing());
-
-    way.addTag("motorcar", "yes");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-
-    way.addTag("bicycle", "yes");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.BICYCLE_AND_CAR));
-
-    way.addTag("foot", "yes");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-  }
-
-  /**
-   * Tests that motorcar/bicycle/foot private don't add permissions but no remove permission if
-   * access is yes
-   */
-  @Test
-  void testMotorCarTagDeniedPermissions() {
-    OSMWay way = new OSMWay();
-    way.addTag("highway", "residential");
-    var permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-
-    way.addTag("motorcar", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.PEDESTRIAN_AND_BICYCLE));
-
-    way.addTag("bicycle", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.PEDESTRIAN));
-
-    way.addTag("foot", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allowsNothing());
-    //normal road with specific mode of transport private only is doubtful
-    /*way.addTag("motorcar", "private");
-        way.addTag("bicycle", "private");
-        way.addTag("foot", "private");
-        permissionPair = getWayProperties(way);
-        assertTrue(permissionPair.main().allowsNothing());*/
-  }
-
-  /**
-   * Tests that motor_vehicle/bicycle/foot private don't add permissions but yes add permission if
-   * access is no
-   * <p>
-   * Support for motor_vehicle was added in #1881
-   */
-  @Test
-  void testMotorVehicleTagAllowedPermissions() {
-    OSMWay way = new OSMWay();
-    way.addTag("highway", "residential");
-    var permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-
-    way.addTag("access", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allowsNothing());
-
-    way.addTag("motor_vehicle", "private");
-    way.addTag("bicycle", "private");
-    way.addTag("foot", "private");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allowsNothing());
-
-    way.addTag("motor_vehicle", "yes");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-
-    way.addTag("bicycle", "yes");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.BICYCLE_AND_CAR));
-
-    way.addTag("foot", "yes");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-  }
-
-  /**
-   * Tests that motor_vehicle/bicycle/foot private don't add permissions but no remove permission if
-   * access is yes
-   * <p>
-   * Support for motor_vehicle was added in #1881
-   */
-  @Test
-  void testMotorVehicleTagDeniedPermissions() {
-    OSMWay way = new OSMWay();
-    way.addTag("highway", "residential");
-    var permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.ALL));
-
-    way.addTag("motor_vehicle", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.PEDESTRIAN_AND_BICYCLE));
-
-    way.addTag("bicycle", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.PEDESTRIAN));
-
-    way.addTag("foot", "no");
-    permissionPair = getWayProperties(way);
-    assertTrue(permissionPair.main().allowsNothing());
-    //normal road with specific mode of transport private only is doubtful
-    /*way.addTag("motor_vehicle", "private");
-        way.addTag("bicycle", "private");
-        way.addTag("foot", "private");
-        permissionPair = getWayProperties(way);
-        assertTrue(permissionPair.main().allowsNothing());*/
-  }
-
-  @Test
-  void testSidepathPermissions() {
-    OSMWay way = new OSMWay();
-    way.addTag("bicycle", "use_sidepath");
-    way.addTag("highway", "primary");
-    way.addTag("lanes", "2");
-    way.addTag("maxspeed", "70");
-    way.addTag("oneway", "yes");
-    var permissionPair = getWayProperties(way);
-
-    assertFalse(permissionPair.main().allows(StreetTraversalPermission.BICYCLE));
-    assertFalse(permissionPair.back().allows(StreetTraversalPermission.BICYCLE));
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-    assertFalse(permissionPair.back().allows(StreetTraversalPermission.CAR));
-
-    way = new OSMWay();
-    way.addTag("bicycle:forward", "use_sidepath");
-    way.addTag("highway", "tertiary");
-    permissionPair = getWayProperties(way);
-
-    assertFalse(permissionPair.main().allows(StreetTraversalPermission.BICYCLE));
-    assertTrue(permissionPair.back().allows(StreetTraversalPermission.BICYCLE));
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-    assertTrue(permissionPair.back().allows(StreetTraversalPermission.CAR));
-
-    way = new OSMWay();
-    way.addTag("bicycle:backward", "use_sidepath");
-    way.addTag("highway", "tertiary");
-    permissionPair = getWayProperties(way);
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.BICYCLE));
-    assertFalse(permissionPair.back().allows(StreetTraversalPermission.BICYCLE));
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-    assertTrue(permissionPair.back().allows(StreetTraversalPermission.CAR));
-
-    way = new OSMWay();
-    way.addTag("highway", "tertiary");
-    way.addTag("oneway", "yes");
-    way.addTag("oneway:bicycle", "no");
-    permissionPair = getWayProperties(way);
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.BICYCLE));
-    assertTrue(permissionPair.back().allows(StreetTraversalPermission.BICYCLE));
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-    assertFalse(permissionPair.back().allows(StreetTraversalPermission.CAR));
-
-    way.addTag("bicycle:forward", "use_sidepath");
-    permissionPair = getWayProperties(way);
-    assertFalse(permissionPair.main().allows(StreetTraversalPermission.BICYCLE));
-    assertTrue(permissionPair.back().allows(StreetTraversalPermission.BICYCLE));
-
-    assertTrue(permissionPair.main().allows(StreetTraversalPermission.CAR));
-    assertFalse(permissionPair.back().allows(StreetTraversalPermission.CAR));
-  }
-
   @Test
   void escalator() {
     assertFalse(WayTestData.cycleway().isEscalator());
@@ -352,13 +143,5 @@ public class OSMWayTest {
 
     escalator.addTag("conveying", "whoknows?");
     assertFalse(escalator.isEscalator());
-  }
-
-  private StreetTraversalPermissionPair getWayProperties(OSMWay way) {
-    WayPropertySet wayPropertySet = new WayPropertySet();
-    WayProperties wayData = wayPropertySet.getDataForWay(way);
-
-    StreetTraversalPermission def = wayData.getPermission();
-    return way.splitPermissions(def);
   }
 }
